add module to centrally track Ceph release info for the backend
Add a small module with a static hash that includes most relevant
meta-info of ceph release. Ceph 15.2 Octopus was used as rather
arbitrary cut-off.
Not all meta-info will be used, but those basic things seemed to not
hurt to add already now.

Further add some helper to query available and available+supported
release codenames and the default release codename and a getter for a
whole release entry.

We currently to not clone entries when returning them, so callers can
mess with the references content, but as this seems unlikely to cause
fallout for this very specific info lets keep it simple for now.

As a implementation details the module experiments with opting into
modern perl feature set through `use v5.36`, allowing signatures and
avoiding the need for enabling warnings/strict modes.
In general all but one of the features that get enabled are pretty
much safe, the problematic one being the feature_bitwise one, which
separates using bitwise operators for numerical and string data types,
causing potential fallout if one used them for a bitwise operation on
a string that should be interpreted as number.

The module will be used in the next commits.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>