triggers: add path-based trigger interest
to avoid the need to mark every package shipping PVE-related perl code as
activating the explicit trigger. the explicit trigger can still be used for
packages that need to reload the API without shipping a perl module

the explicit trigger activation can be dropped in PVE 9.0 in packages that ship
perl code, without a need for versioned dependencies.

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>