ui: resource tree: fix order of types
when ordering items on the same level, we often sort first by type, but
do this via string comparison of the types.

this leads to weird results if we have e.g. lxc/node/qemu mixed in one
list, as nodes would sort in between lxc and qemu

this patch solves that issue by introducing a fixed order for types, so
we have direct control over the order when mixing types without having
to change the type names.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>