ui: view selector: prepare filterfn to be more efficient
instead of just using the filterfn directly, return a function from
`getFilterFn`. This way, when we want to filter different things
depending on e.g. local storage values, we don't have to do that for
every element, but only once and return the finished filterFn.

while at it, rename 'filterfn' to camel case 'filterFn'.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>