ui: tree: show tooltips for all entries that have sensible information
instead of filtering pools and groups out manually, simply accumulate
the info we want to to have (status/hastate/etc.) and if any of those
exist, show the tooltip.

This results in the nodes also having a tooltip, including their hastate
(such as online/maintenance), and automatically would show such things
in the future if we add those fields to other entries.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>