ui: ceph: don't set default size/min-size when editing existing pool
As else we trigger a change event for the size field which triggers
the sizeChange callback that then re-calculates the default size
suggestion, which might be lower as the value the user configured.

This was reported in the Forum for a 5/4 size/min-size configuration
that got reset to 5/3 on edit.

Report: https://forum.proxmox.com/threads/158798/
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>