fix #6138: ui: storage content: use locale for sorting
by using 'localeCompare'.

Assume the files 'a', 'B' and 'c'. The current default sorting results
in :

'B'
'a'
'c'

(C collation)

With this patch we use locale dependent sorting, so the browser/client
setting controls how it's sorted. For example with english, the list
becomes:

'a'
'B'
'c'

Which makes it easier to visually search for specific entries.
Using the default settings here makes sense (so no explicit
'sensitivity' setting for example.), since that is whats most
expected per locale.

Also enable 'numeric' sorting, so 'foo10' is sorted after 'foo2' as
suggested by Fiona.

Interestingly, the initial sorter (via the 'sorters' property of the
store), cannot seem to reuse the sorter defined in the columns, so we
have to specify it twice, once for the initial sort and once when the
user clicks the column header.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>