ui: lxc create: do not suggest nesting is enabled for privileged CTs
The current implementation is slightly misleading. When creating a
privileged container, the nesting checkbox is disabled but keeps its
current state. However, nesting is not enabled for privileged
containers even if the checkbox was set to true.
With this change, the nesting checkbox gets unticked whenever it is
disabled, making clear that the nesting feature will not be enabled.

Signed-off-by: Michael Köppl <m.koeppl@proxmox.com>
 [TL: slight rewording of the commit message]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>