report: add WWN column to lsblk output
WWNs (WWIDs) are useful information for debugging setups where
multipath is not (correctly) set up, as the WWN reveals which block
devices point to the same LUN. The report currently does not provide
WWNs in such scenarios. There is a symlink for the WWN under
/dev/disk/by-id/wwn-*, but if there are multiple block devices with
the same WWN, the symlink just points to one of them.

Hence, add the WWN column to the lsblk output.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>