ui: fix regression with checking if volume is QEMU backup
The function was changed but not all call sites where adapted,
resulting in a JavaScript exception when, e.g., trying to open the
file-restore dialogue for a backup from PBS storage.

Make the check function generic and support passing either a full
volume info object or the volume ID and the volume format as two
separate string parameters.

Fixes: 3f8246030 ("ui: backup: also check for backup subtype to classify archive")
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>