ceph versions: include raw string in error and avoid shadowing variables
Reusing the variable here for different things just adds unnecessary
error potential, avoid that and use the variable names to encode what
we expect to be in them.

Also include the raw string we try to parse as JSON in the warning, so
that we can have an idea about what went wrong if parsing fails.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>