ui: backup: move change detection help tooltip behind field
instead of having it in a separate line. This could be confusing on
first glance, e.g. if it belongs to the line above or below.

To make it more clear, put it in line with the field itself.

Because this uses a FieldContainer now, we can use that to hide and show
both at the same time, reducing the amount of code a bit.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>