ui: login: display error messages when logging
This allows to distinguish, for example, the following three cases:

 - The pvedaemon service is not responding (595)
 - The pveproxy service is not responding (no error status code)
 - The authentication credentials are not correct (401)

Since different combinations of wrong password and/or wrong user all
report the same error message: "authentication failure (401)" this does
not leak login information.

Another consideration is that the code `resp.status` does not match the
HTTP code seen in the API (a failed authentication has a return code 200
from the point of view of the browser) and its information is lost
without this change.

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Link: https://lore.proxmox.com/20250424134423.364697-1-m.sandoval@proxmox.com
 [TL: resolve merge conflict due to code base reformatting]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>