d/control: run install targed as (fake)root to allow setting uid/gid
We want to install some directories as www-data but as since dpkg
newer than 1.22.13 dpkg-buildpackage does not defaults to using
fakeroot anymore, as most packages do not require that anyway and it
tends to hide issues. [0]

Instead of reverting to the previous default of Rules-Requires-Root
equaling "binary-targets" limit the usage of fakeroot purely to the
dh_auto_install rule by using the "debhelper/upstream-make-install"
keyword.

[0]: man deb-src-control
[1]: /usr/share/doc/dpkg/spec/rootless-builds.txt

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>