ui: backup job edit: move notification related settings to separate tab
The notification settings in the 'General' tab were unfortunately a
source of regular confusion for many people. This was primarily due to
the behavior of the 'notification mode'. The notification mode can
one of the following:
  - notification-system: Emit a notification event to the global
    notification system, where it can be matched on by notification
    matchers and then sent to one or more targets.
  - legacy-sendmail: Old-style notifications, where one can directly
    enter some email address. The system uses 'sendmail' to
    send the notification to the specified address, circumventing
    the regular notification stack.
  - auto: Use legacy-sendmail if an email is entered and the
    notification system if not

The 'auto' mode was originally intended to ease migration between the
old and the new system. From a user's perspective however, 'auto' is
quite surprising and unintuitive. The UI preselected 'auto' as a
default, which would, as explained above, favor the new notification
stack with the 'mailto' field empty. However, the UI would still invite
the user to enter their email address, which would then entail the
'legacy-sendmail' mode. Some users were led to believe that this email
address would then be used for a configured email target of the new
notification stack. As a consequence, 'auto' is now completely hidden in
the UI.

In the new 'Notifications' tab one can now choose between
  ( ) Use global notification settings
  (x) Use sendmail to send an email
      Recipients: [              ]
      When:       [Always/On Error]

'Recipients' and 'When' are disabled if the first radio box is selected.

The new tab can later also be used to house other controls. For example,
we could display all matchers that could potentially match this backup
job, or maybe even allow to create a new matcher with a pre-populated
match-field rule.

The term 'Notification System' is altogether from the UI. It is not
necessarily clear to a user that this refers to the settings in
Datacenter > Notifications.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Michael Köppl <m.koeppl@proxmox.com>
Reviewed-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250624120032.153539-1-l.wagner@proxmox.com