ui: oneshot backup: stop shifting UI when selecting 'legacy-sendmail'
The 'protected' checkbox in the left column seems to use slightly less
space than the 'mailto' field in the same row in the right column.
The 'mailto' field is only shown when the 'notification-mode' is set to
legacy-sendmail. Due to the differences in size, the UI shifts by a
couple pixels when the additional field is shown.

As a workaround, the checkbox is padded on the bottom by a tiny amount,
stopping the UI from shifting around when the additional field is
shown.

Signed-off-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20250624120032.153539-3-l.wagner@proxmox.com