8to9: update check for NVIDIA vgpu mgr to newer systemctl is-enabled behavior
systemctl is-enabled now returns an explicit `not-found` instead of a
message on stderr and exit code, we need to adapt to that. Doing so in
the unit state query helper directly is rather wrong as we also need
to support the systemd from PVE 8, and even besides that it might be
better to not conflate an error of that command with the new explicit
not-found state for a future 9to10, the error could be totally
unrelated (e.g., OOM kill) and thus should be differentiated in the
future.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>