fix #5894: pvestatd: improve broadcast of node version-info
Until now, the pvestatd did broadcast the pve-manager version only
once after startup of the service. But there are some situations,
where the local pmxcfs (pve-cluster) restarts and loses that
information.
Basically every time we restarted the pmxcfs without restarting
pvestatd too.

While updates involving the pve-cluster packages normally correlate
with updates of libpve-cluster-api-perl, which cause a restart of all
perl based daemons through the "pve-api-updates" package trigger, it
still could happen in other situations, like for example, on a cluster
join, or if the pmxcfs has been restarted manually.

By additionally checking if the local kv-store of the pmxcfs has any
version info for the node, we can decide if another broadcast is
necessary.

Therefore after the next run of pvestatd, which normally happens every
10s, we should have the full version info available again.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Link: https://lore.proxmox.com/20250714082358.56826-1-a.lauterer@proxmox.com
 [TL: mention packaging triggers]