api/ui: show/return alternative interface names
in api listing for all interfaces, and in the GET call for an individual
interface.

For intefaces that are already using an 'altname' in the iface stanza,
we look up the 'legacy' interface name and the remaining altnames and
show those.

If we don't show them, the user might be confused if the bridge ports
and interface names don't correlate.

enables the additional alternative names column in the network view on
the gui.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250715090749.1608768-3-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>