guest metrics: support and prefer new pve-{type}-9.0 RRD files
We add a new function to handle different key names, as it would
otherwise become quite unreadable.

It checks which key format exists for the type and resource:
* the old pve2-{type} / pve2.3-vm
* the new pve-{type}-{version}

and will return the one that was found. Since we will only have one key
per resource, we can return on the first hit.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>
Link: https://lore.proxmox.com/20250715143218.1548306-16-a.lauterer@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>